+++ src/types/FilteredStatus.ts
@@ -1,0 +1,1 @@
+export type FilteredStatus = 'all' | 'active' | 'completed';
